fix(release): restore workspace and Windows builds
